refactor(storage): extract AgentPlaybookStoreMixin from PlaybookMixin
Move the agent-playbook CRUD + search bucket (18 public methods + the
sqlite-private _index_agent_playbook_fts_vec / _insert_agent_playbook_row)
out of PlaybookMixin into a new AgentPlaybookStoreMixin in the existing
playbook/ package, for both the sqlite concrete backend and the storage_base
ABC. Behavior-preserving verbatim move: method bodies are byte-identical and
the mutation -> rowcount-guarded emit -> commit-in-lock -> self-committing
FTS/vec-after-commit ordering is unchanged.

Shared helpers (_emit_hard_delete_playbook, _build_tags_sql, AGGREGATE_REASON_PREFIX,
_delete_playbook_search_rows) stay on their existing modules and are imported;
the agent-only _emit_supersede_playbook / _AGENT_PLAYBOOK_DEFAULT_EXCLUDED_STATUSES
(sqlite) and _AGGREGATE_EVENT_EMIT_ATTEMPTS (ABC) move with the bucket. Compose
AgentPlaybookStoreMixin into SQLiteStorage and BaseStorage alongside
UserPlaybookStoreMixin. The atomicity/crash-window tests repoint their module
patch targets to playbook._agent to follow the relocated symbols.
