+++ reflexio/server/services/storage/storage_base/playbook/__init__.py
@@ -1,3 +1,4 @@
+from ._agent import AgentPlaybookStoreMixin
 from ._user import UserPlaybookStoreMixin
 
-__all__ = ["UserPlaybookStoreMixin"]
+__all__ = ["AgentPlaybookStoreMixin", "UserPlaybookStoreMixin"]
